Rewrite event handler as separate method
It *was* a TypeScript problem, `$event.target.value` was the
right syntax. But, TS insisted on handling nulls and such
properly, which is much easier to do in the explicit method
than in the inline event handler function inside the <select>
itself.

There might be a way to wrap all of this into an inline function,
but it's probably hard to read and smelly.

The `as` may be required in order to cast the generic Event
into a type that's actually got all the properties of interest.
Loosely similar to catching an Exception that's actually some
subtype of Exception with extra members. Python will duck-type;
*mypy* probably *wouldn't* in this case. The `as` gives the typing
system insight into the real type that's at play, which the system
is not able to infer automatically.